pyproject.toml
[project]
name = "clashking-api"
version = "0.1.0"
description = "API serving Clash of Clans stats for ClashKing"

requires-python = "==3.13.7"
dependencies = [
    "aiocache",
    "aiohttp",
    "bcrypt",
    "coc-py @ git+https://github.com/mathsman5133/coc.py.git@march-update",
    "cryptography",
    "expiring-dict",
    "fastapi",
    "fastapi-mail",
    "gunicorn",
    "hashids",
    "hikari @ git+https://github.com/hikari-py/hikari.git",
    "linkd",
    "matplotlib",
    "openpyxl",
    "orjson",
    "pandas",
    "passlib",
    "pendulum",
    "pillow",
    "pydantic>=2.10.3",
    "pyjwt",
    "pymongo>=4.5.0",
    "python-dotenv",
    "python-snappy",
    "redis",
    "requests",
    "sentry-sdk",
    "slowapi",
    "uvloop",
    "uvicorn",
    "fastapi-cache2>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
]

[tool.uv]
package = false
